cloud_functions/requirements_mysql.txt
google-cloud-secret-manager==2.0.0
PyMySQL>=1.0.2
SQLAlchemy>=1.4.44
pandas>=1.5.2
google-cloud-storage>=2.6.0